refactor(api): rename files and move static map interface to game api folder
