+++ src/main/java/adventure_game/Demo.java
@@ -1,13 +1,13 @@
 package adventure_game;
 
 import adventure_game.data_access.InMemoryUserDataAccessObject;
-import adventure_game.data_access.StaticMapDataAccessObject;
+import adventure_game.game_api.GeoapifyStaticMap;
 import adventure_game.entity.AdventureGame;
 import adventure_game.entity.Direction;
 import adventure_game.entity.Location;
 import adventure_game.interface_adapter.move.MoveController;
 import adventure_game.interface_adapter.move.MovePresenter;
-import adventure_game.interface_adapter.move.MoveStaticMapDataAccessInterface;
+import adventure_game.game_api.MoveStaticMapInterface;
 import adventure_game.interface_adapter.move.MoveViewModel;
 import adventure_game.use_case.move.MoveGameDataAccessInterface;
 import adventure_game.use_case.move.MoveInputBoundary;
@@ -25,7 +25,7 @@ public class Demo {
 
 
         MoveGameDataAccessInterface moveGameDataAccess = new InMemoryUserDataAccessObject();
-        MoveStaticMapDataAccessInterface mapService = new StaticMapDataAccessObject();
+        MoveStaticMapInterface mapService = new GeoapifyStaticMap();
 
         MoveViewModel moveViewModel = new MoveViewModel();
         MovePresenter movePresenter = new MovePresenter(moveViewModel, mapService);
