+++ src/main/java/adventure_game/game_api/MoveStaticMapInterface.java
@@ -1,0 +1,7 @@
+package adventure_game.game_api;
+
+import javax.swing.*;
+
+public interface MoveStaticMapInterface {
+    ImageIcon getMapImage(double latitude, double longitude);
+}
