--- src/main/java/adventure_game/interface_adapter/move/MoveStaticMapDataAccessInterface.java
@@ -1,7 +1,0 @@
-package adventure_game.interface_adapter.move;
-
-import javax.swing.*;
-
-public interface MoveStaticMapDataAccessInterface {
-    ImageIcon getMapImage(double latitude, double longitude);
-}
